Update step_by_step_MacOSX_install.md
I was presented with a Password prompt in the terminal immediately after step 6. I _think_ it was a prompt for a Docker pass. Docker was running but I did not have an account. After registering I executed step 6 again and added Docker pass and it worked. (FYI it is very possible I am misinterpreting what occurred.)